Restructured AccountTab folders
