+++ lib/screens/account_tab/components/add_new_username.dart
@@ -5,7 +5,7 @@ import 'package:anonaddy/utilities/form_validator.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../../global_providers.dart';
+import '../../../global_providers.dart';
 
 class AddNewUsername extends StatelessWidget {
   final _textEditController = TextEditingController();
